+++ lib/features/auth/presentation/providers/auth_provider.dart
@@ -6,7 +6,11 @@ Nos permite a nosotros llegar a nuestro backend directamente
 DataSource tiene las conexiones e implementaciones necesarias
 */
 
+import 'dart:async';
+
+import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:local_auth/local_auth.dart';
 import 'package:login_mobile/features/auth/infrastructure/infrastructure.dart';
 import 'package:login_mobile/features/shared/infrastructure/inputs/services/key_value_storage_impl.dart';
 import 'package:login_mobile/features/shared/infrastructure/inputs/services/key_value_storage_services.dart';
@@ -28,29 +32,46 @@ final authProvider = StateNotifierProvider<AuthNotifier, AuthState>((ref) {
 class AuthNotifier extends StateNotifier<AuthState> {
   final AuthRepository authRepository;
   final KeyValueStorageService keyValueStorageService;
+  final LocalAuthentication auth = LocalAuthentication();
+  final Function? closeModalCallback;
 
   AuthNotifier({
     required this.authRepository,
     required this.keyValueStorageService,
+    this.closeModalCallback,
   }) : super(AuthState()) {
     checkAuthStatus();
   } //no hace falta mandar nada porque todo son valores opcinonales
   //estas funciones terminan delegando en el repositorio
 
-  Future<void> loginUser(String email, String password) async {
+  Future<void> loginUser(String username, String password,
+      {bool biometric = false, bool authBiometric = false}) async {
     await Future.delayed(const Duration(milliseconds: 500));
-
     try {
-      final user = await authRepository.login(email, password);
-      _setLoggedUser(user);
+      final user = await authRepository.login(username, password);
+      _setUsername(username);
+      if (biometric == false) {
+        _setTempLoggedUser(user);
+      }
+      if (biometric == true && state.user != null) {
+        _setBiometricLoggedUser(user);
+      }
+      if (closeModalCallback != null) {
+        closeModalCallback!();
+      }
     } on CustomError catch (e) {
-      logout(e.message);
+      if (biometric == false) {
+        logout(e.message);
+      } else {
+        biometricError(e.message);
+      }
     } catch (e) {
-      logout('Something went wrong');
+      if (biometric == false) {
+        logout('Something went wrong');
+      } else {
+        biometricError('Your username or password is incorrect');
+      }
     }
-
-    //final user = await authRepository.login(email, password);
-    //state = state.copyWith(user: user, authStatus: AuthStatus.authenticated); //copia del usuario donde ya tengo el estado
   }
 
   //TODO: implementar el registro
@@ -70,61 +91,136 @@ class AuthNotifier extends StateNotifier<AuthState> {
   void checkAuthStatus() async {
     final token = await keyValueStorageService.getValue<String>('token');
     if (token == null) return logout();
-
     try {
-      //el repositorio hace la auth
       final user = await authRepository.checkAuthStatus(token);
-      _setLoggedUser(user);
+      _setBiometricLoggedUser(user);
     } catch (e) {
       logout();
     }
   }
 
-  void _setLoggedUser(User user) async {
-    await keyValueStorageService.setKeyValue('token', user.token);
-
+  void _setTempLoggedUser(User user) async {
     state = state.copyWith(
         user: user, errorMessage: '', authStatus: AuthStatus.authenticated);
   }
 
-  void _setRegisteredUser(User user) async {
+  Future<void> _setBiometricLoggedUser(User user) async {
+    await keyValueStorageService.setKeyValue('token', user.token);
+    await keyValueStorageService.setKeyValue('hasBiometric', true);
     state = state.copyWith(
-        user: user, errorMessage: '', authStatus: AuthStatus.registered);
+      user: user,
+      errorMessage: '',
+      authStatus: AuthStatus.authenticated,
+      hasBiometric: true,
+    );
   }
 
-  noRegister([String? errorMessage]) {
+  void disableBiometric(ref) async {
+    await keyValueStorageService.removeKey('token');
+    await keyValueStorageService.removeKey('hasBiometric');
     state = state.copyWith(
         user: null,
-        errorMessage: errorMessage,
-        authStatus: AuthStatus.notRegister);
+        errorMessage: '',
+        authStatus: AuthStatus.authenticated,
+        hasBiometric: false);
+    //Aqui puedo abrir otro modal
+    // showModalBottomSheet(
+    //     isScrollControlled: true,
+    //     context: ref,
+    //     builder: (context) => const CustomLogin());
+  }
+
+  void _setUsername(String username) async {
+    await keyValueStorageService.setKeyValue('username', username);
   }
 
   Future<void> logout([String? errorMessage]) async {
-    await keyValueStorageService.removeKey('token');
+    //await keyValueStorageService.removeKey('token');
+    await keyValueStorageService.removeKey('username');
     state = state.copyWith(
         user: null,
         errorMessage: errorMessage,
         authStatus: AuthStatus.notAuthenticated);
   }
+
+  Future<void> biometricError([String? errorMessage]) async {
+    state = state.copyWith(
+      errorMessage: errorMessage,
+    );
+  }
+
+  Future<bool> authWithBiometrics() async {
+    bool authenticated = false;
+    try {
+      state = state.copyWith(authStatus: AuthStatus.checking);
+      authenticated = await auth.authenticate(
+        localizedReason:
+            'Scan your fingerprint (or face or whatever) to authenticate',
+        options: const AuthenticationOptions(
+          stickyAuth: true,
+          biometricOnly: true,
+        ),
+      );
+      if (authenticated) {
+        state = state.copyWith(authStatus: AuthStatus.authenticated);
+      } else {
+        state = state.copyWith(authStatus: AuthStatus.notAuthenticated);
+      }
+    } on PlatformException catch (e) {
+      state = state.copyWith(
+          errorMessage: 'Error - ${e.message}',
+          authStatus: AuthStatus.notAuthenticated);
+    }
+    return authenticated;
+  }
+
+  Future<bool> loginWithBiometrics() async {
+    try {
+      final token = await keyValueStorageService.getValue<String>('token');
+      if (token == null) {
+        return false;
+      }
+      final user = await authRepository.checkAuthStatus(token);
+      state = state.copyWith(
+        user: user,
+        errorMessage: '',
+        authStatus: AuthStatus.authenticated,
+      );
+      authWithBiometrics();
+      return true;
+    } catch (e) {
+      state = state.copyWith(
+        errorMessage: 'Error: $e',
+        authStatus: AuthStatus.notAuthenticated,
+      );
+      return false;
+    }
+  }
 }
 
-enum AuthStatus { checking, authenticated, notAuthenticated, registered, notRegister }
+enum AuthStatus { checking, authenticated, notAuthenticated, hasBiometric }
 
 class AuthState {
   final AuthStatus authStatus;
   final User? user;
   final String errorMessage;
+  final bool? hasBiometric;
 
   AuthState(
       {this.authStatus = AuthStatus.checking,
       this.user,
-      this.errorMessage = ''});
+      this.errorMessage = '',
+      this.hasBiometric});
 
   AuthState copyWith(
-          {AuthStatus? authStatus, User? user, String? errorMessage}) =>
+          {AuthStatus? authStatus,
+          User? user,
+          String? errorMessage,
+          bool? hasBiometric}) =>
       AuthState(
         authStatus: authStatus ?? this.authStatus,
         user: user ?? this.user,
         errorMessage: errorMessage ?? this.errorMessage,
+        hasBiometric: hasBiometric ?? this.hasBiometric,
       );
 }