+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 0.0.11 (2023-10-04)
+
+### 0.0.10 (2023-10-04)
+
+
+### Bug Fixes
+
+* showSnackbar visible in the showModalBottomSheet and error is visible here ([#21](https://github.com/hawks-atlanta/frontend-flutter/issues/21)) ([fd649f1](https://github.com/hawks-atlanta/frontend-flutter/commit/fd649f14a5eb7ab5751daee81c79c61bc1ce9ff2))
+
+### 0.0.9 (2023-10-03)
+
+
+### Features
+
+* added local biometric authentication ([#19](https://github.com/hawks-atlanta/frontend-flutter/issues/19)) ([0218c88](https://github.com/hawks-atlanta/frontend-flutter/commit/0218c88bb68a92507015a640a08a6eb7df9f81c1)), closes [#12](https://github.com/hawks-atlanta/frontend-flutter/issues/12) [#13](https://github.com/hawks-atlanta/frontend-flutter/issues/13) [#17](https://github.com/hawks-atlanta/frontend-flutter/issues/17)
+
+### 0.0.8 (2023-09-24)
+
 ### 0.0.7 (2023-09-24)